feat: Place Cascade 추가, 삭제 확인 알림창 추가

+++ app/src/main/java/org/konkuk/placelist/main/MainActivity.kt
@@ -99,12 +99,23 @@ class MainActivity : AppCompatActivity(), AddPlaceListener {
             }
 
             override fun onSwiped(viewHolder: RecyclerView.ViewHolder, direction: Int) {
-                CoroutineScope(Dispatchers.IO).launch{
-                    db.placesDao().delete(placeAdapter.items[viewHolder.adapterPosition])
-                    withContext(Dispatchers.Main) {
-                        placeAdapter.removeItem(viewHolder.adapterPosition)
+                AlertDialog.Builder(this@MainActivity)
+                    .setTitle("장소 삭제")
+                    .setMessage("장소 내 할 일이 모두 삭제됩니다.\n정말 삭제하시겠어요?")
+                    .setPositiveButton("삭제"){_, _ ->
+                        CoroutineScope(Dispatchers.IO).launch{
+                            db.placesDao().delete(placeAdapter.items[viewHolder.adapterPosition])
+                            withContext(Dispatchers.Main) {
+                                placeAdapter.removeItem(viewHolder.adapterPosition)
+                            }
+                        }
                     }
-                }
+                    .setNegativeButton("취소"){d, _ ->
+                        placeAdapter.refresh()
+                        d.dismiss()
+                    }
+                    .create()
+                    .show()
             }
         }
         val itemTouchHelper = ItemTouchHelper(simpleCallback)
